Add tests to ModalContent
