+++ src/client/src/modals/PlatformRequirementsModal/PlatformRequirementsModal.test.tsx
@@ -1,20 +1,10 @@
-import { fireEvent } from "@testing-library/react";
 import * as React from "react";
 import configureMockStore from "redux-mock-store";
 
 import { addPlatformRequirementsOptions, getInitialState, setOpenModal } from "../../mockData/mockStore";
-import { closeModalAction } from "../../store/navigation/modals/action";
 import { NAVIGATION_MODAL_TYPES } from "../../store/navigation/typeKeys";
 import { renderWithStore } from "../../testUtils";
 import PlatformRequirementsModal from ".";
-import messages from "./messages";
-
-jest.mock("../../store/navigation/modals/action", () => {
-  const closeModalAction = jest.fn(() => ({
-    type: "WTS/navigation/modals/CLOSE_MODALS",
-  }));
-  return { closeModalAction };
-});
 
 jest.mock("./RequirementItem", () => {
   return () => {
@@ -43,14 +33,6 @@ describe("PlatformRequirementsModal", () => {
   it("renders without crashing", () => {
     const wrapper = renderWithStore(<PlatformRequirementsModal {...props} />, store);
     expect(wrapper).toBeDefined();
-  });
-
-  it("On press close button, close modal", () => {
-    const { getByLabelText } = renderWithStore(<PlatformRequirementsModal {...props} />, store);
-    const buttonLabel = intl.formatMessage(messages.ariaCloseModalLabel);
-    const button = getByLabelText(buttonLabel);
-    fireEvent.click(button);
-    expect(closeModalAction).toBeCalled();
   });
 
   it("should be the same number of requirements components as requirements in the store", () => {
